- DTO (Fix) - DAO (Fix) - Service & Repo Fixes & Improvements - Controller & WebController Redesign - Converter Implementation - (Mongo & Jpa) (Repo & Service) works - etc

+++ src/main/java/com/univ/db/service/impl/JpaService.java
@@ -5,6 +5,8 @@ package com.univ.db.service.impl;
  */
 
 import com.univ.db.service.ICRUDService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -17,8 +19,31 @@ public class JpaService<T> implements ICRUDService<T, Long>{
 
     private JpaRepository<T, Long> primeRepository;
 
+    private final Logger LOGGER = LoggerFactory.getLogger(JpaService.class);
+
     public JpaService(JpaRepository<T, Long> jpaRepository) {
         this.primeRepository = jpaRepository;
+    }
+
+    protected boolean invalidModel(T t) {
+        if(t == null) {
+            LOGGER.warn("NULL MODEL");
+            return true;
+        }
+        return false;
+    }
+
+    protected boolean invalidId(Long id) {
+        if(id == null) {
+            LOGGER.warn("NULL ID");
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public Optional<Long> count() {
+        return Optional.of(primeRepository.count());
     }
 
     @Override
@@ -28,22 +53,34 @@ public class JpaService<T> implements ICRUDService<T, Long>{
 
     @Override
     public Optional<T> getById(Long id) {
+        if(invalidId(id))
+            return Optional.empty();
+
         return Optional.of(primeRepository.findOne(id));
     }
 
     @Override
     public Optional<T> save(T t) {
+        if(invalidModel(t))
+            return Optional.empty();
+
         return Optional.of(primeRepository.save(t));
     }
 
     @Override
-    public T delete(T t) {
+    public Optional<T> delete(T t) {
+        if(invalidModel(t))
+            return Optional.empty();
+
         primeRepository.delete(t);
-        return t;
+        return Optional.of(t);
     }
 
     @Override
     public void deleteById(Long id) {
+        if(invalidId(id))
+            return;
+
         primeRepository.delete(id);
     }
 }
